+++ packages/framer-motion/src/animation/animate/__tests__/effects.test.ts
@@ -23,10 +23,8 @@ const subjectEffect = createEffect<Subject>(
             key,
             value,
             () => {
-                subject.values[key] = state.latest[key]
-            },
-            undefined,
-            false
+                subject.values[key] = value.get()
+            }
         ),
     {
         test: (subject): subject is Subject =>